Consolidate library test fixtures
