+++ crates/shelterwood-mailbox/src/lib.rs
@@ -26,6 +26,8 @@ mod deadline;
 mod errors;
 mod futures;
 mod reply;
+#[cfg(test)]
+mod test_support;
 
 #[doc(hidden)]
 pub use capability::{
